Increased the coefficient of friction for the wheels in the constants to match that of the new wheels, moved the vision commands into the vision subfolder, and changed all refrences to visionSubsystem and swerveSubsystem in TrackAprilTagCommand to RobotContainer.[visionSubsystem/swerveSubsystem]
