+++ src/main/deploy/swerve/modules/physicalproperties.json
@@ -1,7 +1,7 @@
 {
   "optimalVoltage": 12,
   "robotMass": 50,
-  "wheelGripCoefficientOfFriction": 1.19,
+  "wheelGripCoefficientOfFriction": 1.43,
   "currentLimit": {
     "drive": 40,
     "angle": 20
