+++ src/main/java/frc/robot/commands/vision/DistanceCalculator.java
@@ -1,14 +1,14 @@
-package frc.robot.commands.drive;
+package frc.robot.commands.vision;
 
 import frc.robot.constants.VisionConstants;
 
 public class DistanceCalculator {
-    public static double getDistance(double area) {
+    public static double getDistance(double yaw, double pitch, double area) {
         /*
         This is written in LaTeX notation.
         If you want to view this math, paste it into a LaTeX editor
          
-        \text{Let $f(d)$ be the area at distance $d$} \\
+        \text{Let $f(d)$ be the area at distance $d$ if the target is perfectly flat (not accounting for yaw and pitch)} \\
         \text{Let $p$ represent the exponent that governs the scaling behaviour of $f(d)$} \\ \\
         f(d) = 10\left(\frac{d}{\text{AREA_10_DISTANCE}}\right)^p \\ \\
         f\left(2\left(\text{AREA_10_DISTANCE}\right)\right) = \sqrt{10} \\ \\
@@ -31,6 +31,7 @@ public class DistanceCalculator {
         d \approx \text{AREA_10_DISTANCE}\left(\frac{\text{area}}{10}\right)^{-\frac{1}{1.6609640474437}} \\ \\
         d \approx \text{AREA_10_DISTANCE}\left(\frac{\text{area}}{10}\right)^{-0.602059991328} \\ \\
         d \approx \text{AREA_10_DISTANCE}\left(\frac{10}{\text{area}}\right)^0.602059991328
+        
 
         In simple terms, d = AREA_10_DISTANCE * (10/area)^0.602059991328
         */
